# Update optimization plugin