--- a/package.json
+++ b/package.json
@@ -97,6 +97,7 @@
     "stylelint-scss": "^3.3.2",
     "stylelint-webpack-plugin": "^0.10.5",
     "svg-spritemap-webpack-plugin": "^3.0.0-rc.2",
+    "terser-webpack-plugin": "^1.1.0",
     "url-loader": "^1.1.2",
     "vue-jest": "^3.0.0",
     "vue-loader": "^15.4.2",