This should fix the extended lag seen at "92% chunk asset optimization UglifyJSPlugin" when compiling an app for prod.

This has been a common annoyance for webpack users lately, due to some unoptimized code for node 6 compatibility deep in uglifyJS
https://github.com/webpack-contrib/uglifyjs-webpack-plugin/issues/272

one of the webpack maintainers recommends using the terser-webpack-plugin, which is another minify/uglify tool that will be the webpack standard in v5, but won't be updated/required for v4 due to some breaking changes.
https://github.com/facebook/create-react-app/issues/4492#issuecomment-421959548

some preliminary testing:


current status, master branch (uglifyJS)
```
build:pl
Version: webpack 4.23.1
Time: 326635ms

build:drupal
Version: webpack 4.23.1
Time: 303973ms
```

updated state, terser branch (terser-webpack-plugin)
```
build:pl
Version: webpack 4.23.1
Time: 34814ms

build:drupal
Version: webpack 4.23.1
Time: 29540ms
```

decrease in PL prod compile time: 291821ms or **4.86m faster**
decrease in drupal prod compile time: 274433ms or **4.57m faster**